README.md
This project is a professional website built using PHP, HTML, Docker, OpenSSL, and various Azure technologies including Keyvaults, App Services, Front Door, and WAF. The website serves as a resume, blog, and portfolio, showcasing the developer's proficiency in web development and design. The project demonstrates expertise in security protocols and secure data transfer, as well as knowledge of SSL certificate creation and management. Additionally, the developer utilized Azure's cloud services and Docker to design and develop the web application, showcasing problem-solving ability and logical reasoning skills. The project also highlights the developer's commitment to continuous learning and innovation, as well as independence and time management skills.